requirements.txt
text
pandas
numpy
scikit-learn
umap-learn
hdbscan
networkx
matplotlib
sentence-transformers
torch
tqdm
Pillow
clip @ git+https://github.com/openai/CLIP.git
